Add optional configurations `external_server_host` and `external_server_port`
Server maintainers often want to run the main aware-micro process as the backend of
reverse proxies (such as Nginx), or load balancers (such as Amazon ALB).

But, the existing configurations `server_host` and `server_port` are used both for
listening as the main aware-micro process, and showing for users/logs as public URLs.

It adds new optional configurations `external_server_host` and `external_server_port`
to split listening and showing.